requirements.txt
# requirements.txt
chromadb
sentence-transformers
ollama
pypdf
python-docx
openpyxl
pyyaml
python-dotenv
tqdm
rich